mgr/volumes: Derive v2 from v1 to leverage common methods
With v2 introduction in subvolumes, there is quite a bit of
common code and methods that both v1 and v2 share.

To reduce overall code bloat and improve maintainability, derive
SubvolumeV2 from SubvolumeV1.

(cherry picked from commit 0b80de9b6a3b950848fde8afa7a0d32b4641d26c)
